requirements.txt
python-dotenv

llama-index-core
llama-index-llms-openai
llama-index-utils-workflow
llama-index-workflows[server]

openinference-instrumentation-llama-index
"opentelemetry-proto>=1.12.0" 
opentelemetry-exporter-otlp 
opentelemetry-sdk 
arize-otel 

tavily-python